// src/app/modules/land-creation/models/land_charge_creation.model.ts
export class LandChargeCreation {
    ChargeCreationID: number;
    LandInfoID: string;
    LegalDocNo: string;
    PassbookNO: string;
    Tehsil: string;
    Province: string;
    District: string;
    Village: string;
    UserID: string;
    BranchId: number;
    Remarks: string;
    ReferenceNo: string;
    IsActive: string;
    TotalArea: string;
}